requirements.txt
numpy                       ==2.1.2
datasets                    >=3.2.0
transformers                ==4.57.6
tokenizers                  >=0.21.0
accelerate                  >=0.26.0
bitsandbytes
evaluate

Pillow
opencv-python
tqdm
nltk
json-repair
